Backend/backend-api/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
psycopg2-binary
bcrypt